auth-ldap: Fix incorrect usage of unconfigured dns
If the DNS servers were not considered, the empty string would be
interpreted as a list of one DNS server with an empty IP address. Therefore,
the name servers specified in the system DNS configuration would not be
used.

This patch corrects the issue by filtering empty DNS server specifications.