Add autoreply filter concept
mail-autoreply is a simple filter which allows an administrator to configure
a list of regular expressions executed against the subject and body of
incoming emails (piped and fetched). If any match, the mail is flagged as an
automated response and will instruct osTicket not to send returning
automated responses.

This is a fair last mile for automated responses which either have no
connection to previous messages or do not use a MIME header to indicate they
are automated.