thread: Allow edits to the attachments of thread entry
This allows administrators with access to a thread and staff members
access to edit the attachments of a thread entry item. Non
administrators can only edit the attachments if they own the thread
entry.

This is helpful if the incorrect attachments were submitted with a
posting, or if the file names should be edited, or if additional
attachments should have been added with the posting.

Edits to responses will not be resent to the end user. Instead, the
posting should be edited afterwards and then resent. Perhaps a note
could be added to the amended responses to indicate that the attachments
were modified.