ldap: PHP 7.2+ Host:Port Conflict
This addresses 160 where setting `host:port` for the "LDAP Servers" option
does not work when using PHP 7.2+. In PHP 7.2+ the `host:port` string will
be used as the full hostname and since we are not passing a port the system
assumes `389`. The end result of this is `host:port:389` which is not valid
and fails every time. This adds a regex check for the port number in the
string and if exists we add the port number to the `port` option in the
`$info` array for `Net_LDAP2::bind()`.  The end result of this is
`host:port` which is valid and will connect successfully every time. If no
port is passed the system assumes the port is the default (`389`) and the
end result will be `host:389`.