audit: New Agent DB Error
This addresses an issue where having the Audit Log plugin Enabled and
creating a new Agent throws a DB Error 1064. This error is due to an
incomplete Staff object (as it's being created) which causes issues in later
methods. This adds a check for a Staff ID and if none we avoid running the
Audit queries.