Issue: Audit Events
This commit fixes an issue where we would always try to add events to the event table when installing the plugin when it was possible that they already existed. Now, we'll check to make sure those events don't exist before creating them, and if they do exist, log a warning (if warnings are being logged for the helpdesk).