Attachments in Amazon S3: Folders
This commit adds a new field to the plugin configuration that allows the Admin to put in an S3 Folder Path if they want to upload files to a specific S3 folder. If no folder is provided, the file will be uploaded directly to the S3 Bucket. If a folder is specified and it does not currently exist in S3, the folder will be created.