Additional S3 Folder Functionality
This commit adds extra functionality needed to ensure that using folders in S3 will still work with helpdesks using the existing plugin without folders as well as working if the name of the S3 folder is changed. We will now store the S3 bucket and folder name in the attrs of the file record so that we know where to retrieve the file from. If no attrs exist and you are downloading a file, we will download from the bucket configured in the plugin