Issue: 2FA URL Spaces
This commit adds an option for an Admin to add a custom Issuer name in the Config for the 2FA plugin. The issueer is the title shown i
n an Authenticator app after the QR code is scanned. If no configuration is set up, we use 'osTicket' as the issuer name..

It also fixes an issue where some authenticators cannot handle spaces or special characters in the URL. It strips out spaces and special characters when generating the URL for the QR code.