issue: Map Custom Dependencies
This adds functionality to map custom decencies on hydrate. A custom
dependency is a decency that is not managed through Composer (like auth-ldap
is now). This means any decency that is included in lib/ and is configured
to be mapped (via plugin config) will be mapped as normal Composer
dependencies.