auth-oauth2: Disable Strict Email Matching on Authorization
This commit disables strict email match when doing email authorization (getting
token). While strict by default was noble, checking for a match is troublesome
when a global admin can authorize for an email account or/and its aliases.

A strict placeholder flag is set to false default for now with the intention of
making it configurable in the future.