ldap: Staff Authentication Backend ID
This addresses an issue where we are still using the static ID instead
of using the Backend ID causing Agent lookups and searches to fail. This
updates both places we use the static ID to `getBkId()` instead. This
will retrieve the proper Backend ID and will allow Agent lookups and
searches to function successfully.