Microsoft OAuth2 - Consent Prompt Option
This provides the option for the user to either enable or disable a consent prompt from Microsoft when authorising the OAuth2 request with Microsoft 365.

This was discussed as a possible option in Merge Request #241, this provides that option to the end user when configuring the OAuth2 endpoint.

The new option is shown in the **IdP Config** tab in the OAuth2 Authorisation popup.