Microsoft OAuth2 - Strict Email Match Option
This provides the option for the user to either enable or disable strict email matching rather than having it hard coded in the plugin.

- When enabled the email address configured in osTicket must match the account authorising the OAuth2 request.
- When disabled the email address cab be authorised by a User Mailbox that has permissions to send and receive on behalf of the email address configured in osTicket, this is a requirement for Shared Mailboxes to be able to send emails as they must be sent from a licenced user.

The fix for Shared Mailboxes in Microsoft 365 is in Merge Request #261.

The new option is shown in the **IdP Config** tab in the OAuth2 Authorisation popup.