issue: resolveDependencies Scripts and Extra
This resolves and issue where when hydrating the S3 plugin it's still
including uneeded dependencies even though the plugin info includes the
script and extra tags. This is due to `make.php` not taking those into
account when resolving the dependencies. This updates the
`resolveDependencies` method to account for the scripts and extra tags
and will include them if found. This also addresses an small issue
where the extra value needs to be an array otherwise it causes fatal
error.